Added Scan Option for new Modules!
